refactor: 모달 닫기와 관련된 기능을 custom hook으로 적용
